ai_project/requirements.txt
asgiref==3.5.0
Django==4.0.4
sqlparse==0.4.2
gunicorn==20.1.0
numpy==1.20.0
joblib==1.1.0
Pillow==9.2.0
tensorflow==2.11.0
keras==2.11.0
pandas==1.4.2
scipy==1.7.3